[MWPW-200314] fix(sessions-guide): dedupe parseConfig, drop stale featured-ids, use async/await in icon-resolver
sessions-guide.js and sessions-guide-full-page.js each had their own copy of
parseConfig() that had already drifted: the full-page variant was missing
the 'featured-sessions' case entirely, so authoring that key there silently
no-opped. Extracts one shared parseSessionsGuideConfig() (utils/parse-config.js),
parameterized by logPrefix and forcedSurface, so a new authoring key can only
be added in one place going forward.

Also removes REAL_FEATURED_IDS, a hardcoded fallback tied to a specific event
date that has already passed. getFeaturedSessions() already falls back to a
deterministic shuffle of real sessions when featuredSessionIds is empty, so
this is a strictly better default than stale IDs from a past event.

Converts icon-resolver.js's .then() chains to async/await (loadMiloIcons/
loadOwnIcons split into a sync memoizing wrapper + a separate async fetch
function, preserving the same promise-caching behavior) and wraps Icon.js's
effect body in an async IIFE, since useEffect callbacks can't be async
directly.
